+++ projeto_final/tests/PathTest.cpp
@@ -78,7 +78,7 @@ TEST(Path, totalSwap){
 
   EXPECT_NEAR(path->cost, 8 + 4*sqrt(2), 1e-6);
 
-  int mini, minj; float minchange;
+  int mini, minj; double minchange;
   findMiniMinjOpenMP(path, &mini, &minj, &minchange);
 
   EXPECT_EQ(1, mini);
